#61: Implement ServiceCollectionExtensions in Core layer
