+++ Api/ShoppingList.Api.Core/ShoppingList.Api.Core.csproj
@@ -7,6 +7,7 @@
 
   <ItemGroup>
     <PackageReference Include="AutoFixture" Version="4.15.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0" />
   </ItemGroup>
 
 </Project>
