+++ Api/ShoppingList.Api.Endpoint/ServiceCollectionExtensions.cs
@@ -1,9 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using ProjectHermes.ShoppingList.Api.Core.Converter;
+using ProjectHermes.ShoppingList.Api.Core.Extensions;
 using ProjectHermes.ShoppingList.Api.Endpoint.v1.Controllers;
-using System;
-using System.Linq;
-using System.Reflection;
 
 namespace ProjectHermes.ShoppingList.Api.Endpoint
 {
@@ -20,42 +18,6 @@ namespace ProjectHermes.ShoppingList.Api.Endpoint
             var assembly = typeof(ServiceCollectionExtensions).Assembly;
             services.AddConverters(assembly, typeof(IToContractConverter<,>));
             services.AddConverters(assembly, typeof(IToDomainConverter<,>));
-        }
-
-        public static void AddConverters(this IServiceCollection services, Assembly assembly,
-            Type converterType)
-        {
-            var assemblyTypes = assembly
-                .GetTypes()
-                .Where(t => !t.IsAbstract)
-                .ToList();
-
-            foreach (var assemblyType in assemblyTypes)
-            {
-                var interfaceTypes = assemblyType
-                    .GetInterfaces()
-                    .Where(type => type.IsGenericType
-                        && type.GetGenericTypeDefinition() == converterType
-                        && services.All(service => !IsTypeIsInDescriptor(service, type, assemblyType)));
-
-                foreach (var interfaceType in interfaceTypes)
-                {
-                    services.AddTransient(interfaceType, assemblyType);
-                }
-            }
-        }
-
-        private static bool IsTypeIsInDescriptor(ServiceDescriptor descriptor, Type serviceType, Type implementationType)
-        {
-            if (descriptor.ServiceType != serviceType)
-                return false;
-
-            if (descriptor.ImplementationType == implementationType
-                || descriptor.ImplementationInstance != null
-                    && descriptor.ImplementationInstance.GetType() == implementationType)
-                return true;
-
-            return false;
         }
     }
 }
